test: target-amount 통합 테스트 파일 통일 및 jdbc bulk query 수정
